pow: only support prime calculations on Go 1.8 and above
